Deal with review comments
The comments were as follows:

**Section 1, second paragraph:

I know that the intention of this paragraph is to describe the state of
the stack when it is done, but currently it uses the present tense
making it seem that all of the features exist in the current stack.
E.g. we do not currently have functioning integrated calibration
products pipelines.  I think it would make it more obvious what you mean
if you change the tense to future for that paragraph.

**Section 1, third paragraph:

It mentions that C++ is used for computationally intensive algorithms,
and that has always been a guiding principle.  However in the current
system, there is a fair amount of infrastructure code in C++ as well
(tables, WCS objects, coordinate transforms, etc.).  It may be closer to
the current implementation to say something like "where the latter is
used for CPU-intensive algorithms, or when the algorithms require access
to complex data structures."  It's up to you.

**Page 3 in the afwTable description:
I think you meant:
detections and measurements --> detection and measurement

**Section 3:

You mention in a footnote that El Capitan is currently not supported.
It's not a major issue, but I wonder if it would be good to mention it
in the text so people don't miss it.