Remove mention of aperture correction
Sandrine made a comment that it was unclear what we meant by aperture
correction in this context. Further discussions with people in DM
clarified that explaining the Summer 2015 aperture correction fixes
to make this accurate would take up more space than was warranted
by the changes themselves so we have simply removed mention in the
paper.